requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
tqdm==4.66.2
